videoToTextApp/VM1/backend/requirements.txt
Flask
moviepy
SpeechRecognition
psycopg2-binary